refactor: wire provider extract modules through search facade
